#515 Some API cleanup when writing documentation

+++ api/BusinessLogic/Categories/CategoryGroupHandler.php
@@ -3,8 +3,12 @@
 namespace BusinessLogic\Categories;
 
 
+use BusinessLogic\Exceptions\AccessViolationException;
+use BusinessLogic\Exceptions\ApiFriendlyException;
+use BusinessLogic\Exceptions\ValidationException;
 use BusinessLogic\Security\UserContext;
 use BusinessLogic\Security\UserPrivilege;
+use BusinessLogic\ValidationModel;
 use DataAccess\Categories\CategoryGroupGateway;
 
 class CategoryGroupHandler extends \BaseClass {
@@ -18,7 +22,13 @@ class CategoryGroupHandler extends \BaseClass {
                                    UserContext $userContext,
                                    $heskSettings) {
         if (!$userContext->admin && !in_array(UserPrivilege::CAN_MANAGE_CATEGORIES, $userContext->permissions)) {
-            throw new \Exception("User {$userContext->id} does not have permission to create category groups!");
+            throw new AccessViolationException("User {$userContext->id} does not have permission to create category groups!");
+        }
+
+        if ($categoryGroup->names === null || count($categoryGroup->names) === 0) {
+            $validationModel = new ValidationModel();
+            $validationModel->errorKeys[] = "NAME_REQUIRED";
+            throw new ValidationException($validationModel);
         }
 
         return $this->categoryGroupGateway->createCategoryGroup($heskSettings, $categoryGroup);
@@ -36,9 +46,28 @@ class CategoryGroupHandler extends \BaseClass {
         }
 
         $this->categoryGroupGateway->updateCategoryGroup($heskSettings, $categoryGroup);
+
+        $categoryGroups = $this->categoryGroupGateway->getAllCategoryGroups($heskSettings);
+        foreach ($categoryGroups as $innerCategoryGroup) {
+            /* @var $innerCategoryGroup CategoryGroup */
+            if ($innerCategoryGroup->id === $categoryGroup->id) {
+                return $innerCategoryGroup;
+            }
+        }
+
+        // Huh? Just return what we're given I guess
+        return $categoryGroup;
     }
 
-    public function deleteCategoryGroup($id, $heskSettings) {
+    public function deleteCategoryGroup($id, $userContext, $heskSettings) {
+        if (!$userContext->admin && !in_array(UserPrivilege::CAN_MANAGE_CATEGORIES, $userContext->permissions)) {
+            throw new AccessViolationException("User {$userContext->id} does not have permission to delete category groups!");
+        }
+
+        if (!$this->categoryGroupGateway->doesCategoryGroupExist($id, $heskSettings)) {
+            throw new ApiFriendlyException("Category group {$id} not found.", "Category Group Not Found", 404);
+        }
+
         $this->categoryGroupGateway->moveCategoriesToParentsParent($id, $heskSettings);
         $this->categoryGroupGateway->deleteCategoryGroup($id, $heskSettings);
     }
